@@ -1,58 +1,27 @@
 name: rich_editor
 description: WYSIWYG editor for Flutter with a rich set of supported formatting options.
-version: 0.0.5
+version: 0.0.6
 homepage: https://github.com/JideGuru/rich_editor
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: '>=2.19.6 <4.0.0'
+  flutter: ">=2.11.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_inappwebview: ^5.3.2
-  mime: ^1.0.0
-  image_picker: ^0.8.3+1
-  flutter_widget_from_html_core: ^0.6.1+1
-  flutter_colorpicker: ^0.5.0
-  path: ^1.8.0
-  xml2json: ^5.3.1
+  flutter_inappwebview: ^5.7.2+3
+  mime: ^1.0.4
+  image_picker: ^0.8.9
+  flutter_widget_from_html_core: ^0.10.1
+  flutter_colorpicker: ^1.0.3
+  path: ^1.8.2
+  xml2json: ^5.3.6
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter.
 flutter:
-  # To add assets to your package, add an assets section, like this:
   assets:
     - assets/editor/
-  #
-  # For details regarding assets in packages, see
-  # https://flutter.dev/assets-and-images/#from-packages
-  #
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
-
-  # To add custom fonts to your package, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts in packages, see
-  # https://flutter.dev/custom-fonts/#from-packages
