@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'html_text.dart';
 
 class HeadingDialog extends StatelessWidget {
-  List formats = [
+  final formats = [
     {'id': '1', 'title': '<h1>Heading 1</h1>'},
     {'id': '2', 'title': '<h2>Heading 2</h2>'},
     {'id': '3', 'title': '<h3>Heading 3</h3>'},
